requirements.txt
azure-mgmt-resource==9.0.0
pytest==5.4.1
